fix braking message type

+++ scripts/joy_teleop.py
@@ -11,7 +11,7 @@ MAX_SPEED = 4
 class JoyController:
     def __init__(self):
         self.vesc_pub = rospy.Publisher("/vesc/ackermann_cmd_mux/input/teleop", AckermannDriveStamped, queue_size=0)
-        self.brake_pub = rospy.Publisher("/vesc/commands/motor/brake", float, queue_size=0)
+        self.brake_pub = rospy.Publisher("/vesc/commands/motor/brake", Float64, queue_size=0)
         self.joy_sub = rospy.Subscriber("/vesc/joy", Joy, self.cmd_cb)
 
     def cmd_cb(self, msg):
